DL_KK2_Chatbot/requirements.txt
google-genai
pypdf
numpy
python-dotenv
